day5/requirements.txt
openai>=1.0.0
pydantic
python-dotenv
termcolor
